Revert "add confirmation modal when slug has been modified"
This reverts commit 639ca9b66112e2ecf75c8ac314b1bbf7b73d4d49.

wip: create generic warning modal component for admin forms

Revert "Revert "add confirmation modal when slug has been modified""

This reverts commit 2c3e64de8adb0ca0196ba3647525bc26b28d8ea6.

create a generic modal to confirm changes in the administration
